Get persisting and instance specific components working again
