--- src/main/java/com/mineinabyss/geary/ecs/components/Component.kt
@@ -1,6 +1,0 @@
-package com.mineinabyss.geary.ecs.components
-
-import com.mineinabyss.geary.ecs.api.autoscan.AutoscanComponent
-
-@AutoscanComponent
-public class Component
