+++ src/main/java/com/mineinabyss/geary/ecs/prefab/PrefabManager.kt
@@ -1,23 +1,22 @@
 package com.mineinabyss.geary.ecs.prefab
 
 import com.mineinabyss.geary.ecs.api.entities.GearyEntity
-import com.mineinabyss.geary.ecs.components.GearyPrefab
 import com.mineinabyss.geary.ecs.components.PrefabKey
 import com.mineinabyss.geary.ecs.prefab.PrefabManager.keys
 import com.uchuhimo.collections.MutableBiMap
 import com.uchuhimo.collections.mutableBiMapOf
 
 /**
- * Manages registered [GearyPrefab]s and accessing them via name, bukkit entity, etc...
+ * Manages registered prefabs and accessing them via name.
  *
- * @property keys A map of [GearyPrefab]s registered with the plugin.
+ * @property keys A list of registered [PrefabKey]s.
  */
 public object PrefabManager {
     public val keys: List<PrefabKey> get() = prefabs.keys.toList()
 
     private val prefabs: MutableBiMap<PrefabKey, GearyEntity> = mutableBiMapOf()
 
-    /** Get a prefab by [name]. */
+    /** Gets a prefab by [name]. */
     public operator fun get(name: PrefabKey): GearyEntity? = prefabs[name]
 
     /** Registers a prefab with Geary. */
@@ -26,7 +25,7 @@ public object PrefabManager {
         prefab.set(name)
     }
 
-    /** Clears all stored [prefabs], but not [persistentTypes] */
+    /** Clears all stored [prefabs] */
     internal fun clear() {
         prefabs.clear()
     }
